feat: remove old API client from project
