+++ tests/conftest.py
@@ -9,7 +9,7 @@ from starlette.testclient import TestClient
 
 from project.config import Settings, MinioBucketConfig, OIDCConfig, HubConfig
 from project.dependencies import get_settings
-from project.hub import AuthWrapper, ApiWrapper
+from project.hub import FlamePasswordAuthClient, FlameHubClient
 from project.server import app
 from tests.common.auth import get_oid_test_jwk
 from tests.common.env import (
@@ -26,7 +26,7 @@ from tests.common.env import (
     PYTEST_HUB_AUTH_PASSWORD,
     PYTEST_HUB_API_BASE_URL,
 )
-from tests.common.helpers import eventually, next_prefixed_name, is_valid_uuid
+from tests.common.helpers import next_prefixed_name
 
 
 def __get_settings_override() -> Settings:
@@ -96,47 +96,66 @@ def rng():
 
 
 @pytest.fixture(scope="package")
-def hub_access_token():
-    return (
-        AuthWrapper(PYTEST_HUB_AUTH_BASE_URL)
-        .acquire_access_token_with_password(
-            PYTEST_HUB_AUTH_USERNAME, PYTEST_HUB_AUTH_PASSWORD
-        )
-        .access_token
+def auth_client():
+    return FlamePasswordAuthClient(
+        PYTEST_HUB_AUTH_USERNAME,
+        PYTEST_HUB_AUTH_PASSWORD,
+        base_url=PYTEST_HUB_AUTH_BASE_URL,
+        force_acquire_on_init=True,
     )
 
 
 @pytest.fixture(scope="package")
-def api(hub_access_token):
-    return ApiWrapper(PYTEST_HUB_API_BASE_URL, hub_access_token)
+def api_client(auth_client):
+    return FlameHubClient(auth_client, base_url=PYTEST_HUB_API_BASE_URL)
 
 
-@pytest.fixture(scope="package")
-def analysis_id(api, rng):
+@pytest.fixture
+def project_id(api_client):
     project_name = next_prefixed_name()
-    project = api.create_project(project_name)
+    project = api_client.create_project(project_name)
 
+    # check that project was successfully created
     assert project.name == project_name
-    assert is_valid_uuid(project.id)
 
+    # check that project can be retrieved
+    project_get = api_client.get_project_by_id(project.id)
+    assert project_get.id == project.id
+
+    # check that project appears in list
+    project_get_list = api_client.get_project_list()
+    assert any([p.id == project.id for p in project_get_list.data])
+
+    yield project.id
+
+    # check that project can be deleted
+    api_client.delete_project(project.id)
+
+    # check that project is no longer found
+    assert api_client.get_project_by_id(project.id) is None
+
+
+@pytest.fixture
+def analysis_id(api_client, project_id):
     analysis_name = next_prefixed_name()
-    analysis = api.create_analysis(analysis_name, project.id)
+    analysis = api_client.create_analysis(analysis_name, project_id)
 
+    # check that analysis was created
     assert analysis.name == analysis_name
-    assert is_valid_uuid(analysis.id)
+    assert analysis.project_id == project_id
 
-    for bucket_type in ("result", "code", "temp"):
+    # check that GET on analysis works
+    analysis_get = api_client.get_analysis_by_id(analysis.id)
+    assert analysis_get.id == analysis.id
 
-        def __bucket_exists():
-            bucket_name = f"analysis-{bucket_type}-files.{analysis.id}"
-            bucket = api.get_bucket(bucket_name)
-
-            if bucket is None:
-                return False
-
-            assert bucket.name == bucket_name
-            assert is_valid_uuid(bucket.id)
-
-        assert eventually(__bucket_exists)
+    # check that analysis appears in list
+    analysis_get_list = api_client.get_analysis_list()
+    assert any([a.id == analysis.id for a in analysis_get_list.data])
 
     yield analysis.id
+
+    # check that DELETE analysis works
+    api_client.delete_analysis(analysis.id)
+
+    # check that analysis is no longer found
+    assert api_client.get_analysis_by_id(analysis.id) is None
